+++ tests/users.rs
@@ -6,7 +6,7 @@ const TEST_USER_ID: &str = include_str!("test_user_id.txt");
 fn user_info() -> errors::Result<()> {
     let cred = credentials::Credentials::from_file("firebase-service-account.json").expect("Read credentials file");
 
-    let user_session = sessions::user::Session::by_user_id(&cred, TEST_USER_ID, false)?;
+    let user_session = UserSession::by_user_id(&cred, TEST_USER_ID, false)?;
 
     println!("users::user_info");
     let user_info_container = users::user_info(&user_session)?;